feat: unit tests added for first login function at my kiva utils
