+++ src/util/myKivaUtils.js
@@ -1,3 +1,7 @@
+import { differenceInMinutes, fromUnixTime } from 'date-fns';
+
+const FIRST_LOGIN_THRESHOLD = 5;
+
 /**
  * Determines whether the provided loan needs a footnote
  *
@@ -18,4 +22,23 @@ export const hasLoanFunFactFootnote = loan => {
 		default:
 			return false;
 	}
+};
+
+/**
+ * Determines whether is first login for the user
+ *
+ * @param lastLogin last login time from token access
+ * @param memberSince member since time from user data
+ * @returns Whether the user is logging in for the first time
+ */
+export const isFirstLogin = (lastLogin, memberSince) => {
+	const lastLoginDate = fromUnixTime(lastLogin);
+	lastLoginDate.setHours(lastLoginDate.getHours() - 1);
+
+	const minutesDiff = differenceInMinutes(
+		lastLoginDate,
+		new Date(memberSince),
+	);
+
+	return minutesDiff < FIRST_LOGIN_THRESHOLD;
 };
